refactor(examples): separate components
